Ignore c mode execution when compiler fails

+++ MindFreak.rb
@@ -429,13 +429,14 @@ if $0 == __FILE__
         file_exe = "#{filename}.exe"
         t = Time.now.to_f
         IO.write(file_c, MindFreak.to_c(program, tape))
-        system("gcc #{file_c} -o #{file_exe} -O2")
-        puts "Compilation time: #{Time.now.to_f - t}s"
-        # Execute
-        t = Time.now.to_f
-        system(file_exe)
-        puts "\nTime: #{Time.now.to_f - t}s"
-        File.delete(file_c, file_exe) unless keep
+        if system("gcc #{file_c} -o #{file_exe} -O2")
+          puts "Compilation time: #{Time.now.to_f - t}s"
+          # Execute
+          t = Time.now.to_f
+          system(file_exe)
+          puts "\nTime: #{Time.now.to_f - t}s"
+          File.delete(file_c, file_exe) unless keep
+        end
       else raise 'Mode not found'
       end
     end
